% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate-sql-base-teradata.R
\name{bool_to_int}
\alias{bool_to_int}
\title{Translatable function to convert boolean to integer}
\usage{
bool_to_int(cond)
}
\arguments{
\item{cond}{condition}
}
\value{
vector. 1 while cond is TRUE, 0 if FALSE
}
\description{
Translatable function to convert boolean to integer
}
