% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate-sql-base-teradata.R
\name{count_if}
\alias{count_if}
\title{Translatable function to count rows satisfied a condition}
\usage{
count_if(cond)
}
\arguments{
\item{cond}{condition}
}
\description{
Translatable function to count rows satisfied a condition
}
