% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{db_list_tables_with_pattern}
\alias{db_list_tables_with_pattern}
\title{List tables with specified pattern}
\usage{
db_list_tables_with_pattern(con, pattern)
}
\arguments{
\item{con}{Teradata connection.}

\item{pattern}{character string containing a regular expression.}
}
\description{
List tables with specified pattern
}
