% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connection.R
\docType{methods}
\name{dbUnQuoteIdentifier,Teradata,SQL-method}
\alias{dbUnQuoteIdentifier,Teradata,SQL-method}
\title{Un-Quote identifiers}
\usage{
\S4method{dbUnQuoteIdentifier}{Teradata,SQL}(conn, x)
}
\arguments{
\item{conn}{A subclass of \linkS4class{DBIConnection}, representing
an active connection to an DBMS.}

\item{x}{A character vector to un-quote.}
}
\description{
Call this method to generate a string that is unquoted. This is the inverse
of \code{DBI::dbQuoteIdentifier}.
}
