% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate-sql-base-teradata.R
\name{like}
\alias{like}
\title{Translatable function for 'LIKE' operator}
\usage{
like(x, pattern)
}
\arguments{
\item{x}{column name}

\item{pattern}{LIKE pattern}
}
\description{
Translatable function for 'LIKE' operator
}
