% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate-sql-base-teradata.R
\name{mod}
\alias{mod}
\title{Translatable Function for '\%\%' operator}
\usage{
mod(x, divisor)
}
\arguments{
\item{x}{dividend}

\item{divisor}{divisor}
}
\value{
modulo
}
\description{
Translatable Function for '\%\%' operator
}
