% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connection.R
\docType{class}
\name{TeradataOdbcConnection}
\alias{TeradataOdbcConnection}
\alias{TeradataOdbcConnection-class}
\alias{show,TeradataOdbcConnection-method}
\alias{dbGetInfo,TeradataOdbcConnection-method}
\title{Teradata Odbc Connection Methods}
\usage{
\S4method{show}{TeradataOdbcConnection}(object)

\S4method{dbGetInfo}{TeradataOdbcConnection}(dbObj, ...)
}
\arguments{
\item{object}{Any R object}

\item{dbObj}{An object inheriting from \linkS4class{DBIObject},
i.e. \linkS4class{DBIDriver}, \linkS4class{DBIConnection},
or a \linkS4class{DBIResult}}

\item{...}{Other arguments to methods.}
}
\description{
Implementations of pure virtual functions defined in the \code{DBI} package
for TeradataOdbcConnection objects.
}
