% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Driver.R
\docType{class}
\name{TeradataOdbcDriver}
\alias{TeradataOdbcDriver}
\alias{TeradataOdbcDriver-class}
\alias{show,TeradataOdbcDriver-method}
\title{Teradata Odbc Driver Methods}
\usage{
\S4method{show}{TeradataOdbcDriver}(object)
}
\arguments{
\item{object}{Any R object}
}
\description{
Implementations of pure virtual functions defined in the \code{DBI} package
for \code{TeradataOdbcDriver} objects.
}
