% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate-sql-base-teradata.R
\name{to_timestamp}
\alias{to_timestamp}
\title{Translatable function to convert UNIX time to time-stamp}
\usage{
to_timestamp(x)
}
\arguments{
\item{x}{column name stored UNIX time (e.g. 1609459200)}
}
\value{
time-stamp (e.g. "2021-01-01 00:00:00")
}
\description{
Translatable function to convert UNIX time to time-stamp
}
