% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src-teradata.R
\name{src_teradata}
\alias{src_teradata}
\alias{tbl.src_teradata}
\title{Connect to Teradata.}
\usage{
src_teradata(dbname = NULL, host = NULL, port = NULL, user = NULL,
  password = NULL, case = FALSE, ...)

\method{tbl}{src_teradata}(src, from, ...)
}
\description{
Connect to Teradata.
}

