% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connection.R
\docType{class}
\name{TeradataOdbcConnection}
\alias{TeradataOdbcConnection}
\alias{TeradataOdbcConnection-class}
\alias{show,TeradataOdbcConnection-method}
\alias{dbQuoteIdentifier,TeradataOdbcConnection,character-method}
\alias{dbListTables,TeradataOdbcConnection-method}
\alias{dbGetInfo,TeradataOdbcConnection-method}
\title{Teradata Odbc Connection Methods}
\usage{
\S4method{show}{TeradataOdbcConnection}(object)

\S4method{dbQuoteIdentifier}{TeradataOdbcConnection,character}(conn, x, ...)

\S4method{dbListTables}{TeradataOdbcConnection}(conn, ...)

\S4method{dbGetInfo}{TeradataOdbcConnection}(dbObj, ...)
}
\arguments{
\item{object}{Any R object}

\item{conn}{A subclass of \linkS4class{DBIConnection}, representing
an active connection to an DBMS.}

\item{x}{A character vector to quote as identifier.}

\item{...}{Other arguments passed on to methods.}

\item{dbObj}{An object inheriting from \linkS4class{DBIObject},
i.e. \linkS4class{DBIDriver}, \linkS4class{DBIConnection},
or a \linkS4class{DBIResult}}
}
\description{
Implementations of pure virtual functions defined in the \code{DBI} package
for TeradataOdbcConnection objects.
}
