% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connection.R
\docType{methods}
\name{TeradataOdbcConnection}
\alias{TeradataOdbcConnection}
\alias{show,Teradata-method}
\alias{dbQuoteIdentifier,Teradata,character-method}
\alias{dbGetInfo,Teradata-method}
\title{Teradata Odbc Connection Methods}
\usage{
\S4method{show}{Teradata}(object)

\S4method{dbQuoteIdentifier}{Teradata,character}(conn, x, ...)

\S4method{dbGetInfo}{Teradata}(dbObj, ...)
}
\arguments{
\item{object}{Any R object}

\item{conn}{A subclass of \linkS4class{DBIConnection}, representing
an active connection to an DBMS.}

\item{x}{A character vector to quote as identifier.}

\item{...}{Other arguments passed on to methods.}

\item{dbObj}{An object inheriting from \linkS4class{DBIObject},
i.e. \linkS4class{DBIDriver}, \linkS4class{DBIConnection},
or a \linkS4class{DBIResult}}
}
\description{
Implementations of pure virtual functions defined in the \code{DBI} package
for \code{Teradata} objects.
}
